requirements.txt
mysql-connector-python==8.2.0
tkcalendar==1.6.1
